# [SPARK-5938][SQL] Generate Row from JSON string efficiently