When we generate `Row` from JSON String, we zip the schema with index for every record and update current row's fields. That is inefficient.

This pr prepares the necessary functions for updating row's fields in advance, and reuses them for every row.

This pr also reuses a `GenericMutableRow`  instead of creating a new row for every record.