@viirya 2015-02-24
cc @marmbrus.