@viirya 2015-02-26
@marmbrus please take a look of this when you have time. Thanks!