@marmbrus 2015-02-27
Not until after the 1.3 release.
On Feb 27, 2015 8:24 AM, "Liang-Chi Hsieh" <notifications@github.com> wrote:

> @marmbrus <https://github.com/marmbrus> do you have time to take a look
> of this?
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/apache/spark/pull/4712#issuecomment-76422284>.
>