@liancheng 2015-02-28
@viirya Could you please add a brief overview of your update in the PR description? It's not only for the sake of code review. PR descriptions are merged into the Git history together with PRs, thus can be greatly helpful when tracking Git history. Thanks!