@viirya 2015-03-17
@marmbrus @liancheng Please take time to review this. Thanks!