@marmbrus 2015-04-03
Sorry for the delay here.  Can you post a performance comparison?

/cc @yhuai