@viirya 2015-04-03
I have tried to compare the performance difference. However, no significant improvement can be found.

Data: 6000 rows of `complexFieldAndType1` in `TestJsonData`. Call `jsonRDD` to create two json RDDs.

old:
29 seconds, 714 milliseconds
29 seconds, 654 milliseconds
29 seconds, 590 milliseconds
29 seconds, 556 milliseconds
29 seconds, 461 milliseconds


new:
29 seconds, 272 milliseconds
29 seconds, 358 milliseconds
28 seconds, 991 milliseconds
29 seconds, 118 milliseconds
28 seconds, 926 milliseconds

Because this pr can't bring much performance improvement, I am closing it now.